# api/src/spec/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Gamified Volunteering API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Refresh access token using a refresh token",
        "parameters": [
          {
            "in": "cookie",
            "name": "refresh_token",
            "required": true,
            "schema": {
              "type": "string",
              "description": "HttpOnly cookie containing the refresh token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "New access token (and new refresh cookie)",
            "headers": {
              "Set-Cookie": {
                "description": "New HttpOnly refresh token cookie",
                "schema": {
                  "type": "string",
                  "example": "refresh_token=...; HttpOnly; Secure; SameSite=Strict; Path=/v1/auth/refresh"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid/expired refresh token or reuse detected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "RefreshResponse": {
        "type": "object",
        "required": [
          "access_token"
        ],
        "properties": {
          "access_token": {
            "type": "string",
            "description": "New JWT, ~15 minutes"
          }
        }
      },
...
      "TokenErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Invalid or expired token"
          }
        }
      },
...
    }
  },
...
}